% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.dis.R
\name{calculate.dis}
\alias{calculate.dis}
\title{calculate pairwise DoR distance between genes}
\usage{
calculate.dis(mat, detection_rate = 0.9, ncore = 1)
}
\arguments{
\item{mat}{the expression matrix.}

\item{ncore}{the number of cores used. Default 1.}

\item{dectection_rate}{the rate set for dectecing distance. If the running time is too long, you can try detection_rate=0.95 or other higher values to filter more genes. Default 0.9.}
}
\value{
an matrix with pairwise DoR distance
}
\description{
calculate pairwise DoR distance between genes, using \code{\link{calculate.dor}}
}
\examples{
## Not run
mat<-as.matrix(read.csv(file="GSM1599494_ES_d0_main.csv",sep=",",header=F,row.names=1))
gene_dis<-calculate.dis(mat=mat,detection_rate=0.9,ncore=4)
## End(Not run)
}
\seealso{
\code{\link{calculate.dor}}
}
