% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate.norm.R
\name{candidate.norm}
\alias{candidate.norm}
\title{normalization by spike candidate}
\usage{
candidate.norm(mat, spike_candidate, ncore = 1)
}
\arguments{
\item{mat}{the expression matrix.}

\item{spike_candidate}{output from \code{\link{dbscan.pick}}}

\item{ncore}{the number of cores used. Default 1.}
}
\value{
a list containing the normalization results for each candidate set
}
\description{
calculate pairwise DoR distance between genes, using \code{\link{calculate.dor}}
}
\examples{
## Not run
candidate_res<-candidate.norm(mat=mat,spike_candidate=spike_candidate,ncore=4)
## End(Not run)
}
\seealso{
candidate_res
}
