% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.count.R
\name{cluster.count}
\alias{cluster.count}
\title{count the size of a cluster}
\usage{
cluster.count(x)
}
\arguments{
\item{x}{a cluster for count}
}
\value{
size of the cluster
}
\description{
count cluster size, used for \code{\link{dbscan.pick}}
}
\examples{
## Not run
##TBD
## End(Not run)
}
\seealso{
\code{\link{dbscan.pick}}
}
