% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.pick.R
\name{dbscan.pick}
\alias{dbscan.pick}
\title{pick internal spike-in genes}
\usage{
dbscan.pick(dis, ngene = (1:floor(nrow(dis)/25)) * 5, solution = 100)
}
\arguments{
\item{dis}{gene distance from \code{\link{calculate.dis}}.}

\item{ngene}{a series of expected number of IS genes.}

\item{solution}{specify the increasing rate of scanning radius. Default 100.}
}
\value{
internal spike-in genes picked up
}
\description{
pick internal spike-in genes, using \code{\link{cluster.count}}.
}
\examples{
## Not run
spike_candidate<-dbscan.pick(dis=gene_dis,ngene=(1:floor(nrow(gene_dis)/25))*5,solution=100)
## End(Not run)
}
\seealso{
\code{\link{cluster.count}} and \code{\link{calculate.dis}}
}
