% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.ref.R
\name{estimate.ref}
\alias{estimate.ref}
\title{calculate reference expression given internal spike-in genes}
\usage{
estimate.ref(mat, ref_gene)
}
\arguments{
\item{mat}{the expression matrix.}

\item{ref_gene}{the location of spike-in genes in the dataset mat}
}
\value{
estimated expression data fot spike-in genes
}
\description{
calculate reference expression given internal spike-in genes
}
\examples{
## Not run
## TBD
## End(Not run)
}
