% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.sf.R
\name{estimate.sf}
\alias{estimate.sf}
\title{calculate size factor for one cell}
\usage{
estimate.sf(cell, ref_expr)
}
\arguments{
\item{cell}{cell data}

\item{ref_expr}{a reference status for cell data}
}
\value{
estimated size factor for one cell
}
\description{
calculate size factor for one cell
}
\examples{
## Not run
## TBD
## End(Not run)
}
