% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.candidate.R
\name{opt.candidate}
\alias{opt.candidate}
\title{Choose an appropriate IS geneset}
\usage{
opt.candidate(mat, candidate_res, baseline_threshold = 0.1,
  p_value = 0.05)
}
\arguments{
\item{mat}{the expression matrix.}

\item{candidate_res}{output from \code{\link{candidate.norm}}}

\item{baseline_threshold}{specifies the threshold of instability score used to choose the baseline geneset (see our article for more details).}

\item{p_value}{specifies the threshold of p value for F-test. Default 0.05.}
}
\value{
a list with 5 elements.The output ISnorm_res$normalized contains the normalized matrix. The output ISnorm_res$size_factor contains the size factor for each cell. These two outputs can be used as inputs for other scRNA-seq analysis. The output ISnorm_res$ISgenes contains the name of IS genes used for normalization. The output ISnorm_res$inst_cell contains the instability score for each cell. The output ISnorm_res$picked contains the index of the optimized candidate geneset. A message will also tell you the index of the chosen geneset.
}
\description{
Choose an appropriate IS geneset based on the results from candidate_res.
}
\examples{
## Not run
candidate_res<-candidate.norm(mat=mat,spike_candidate=spike_candidate,ncore=4)
## End(Not run)
}
\seealso{
\code{\link{candidate.norm}}
}
